refactor: Optimized folder structure
