+++ public/register.html
@@ -4,7 +4,7 @@
 		<meta charset="utf-8">
 		<title>Login Form Tutorial</title>
 		<link href="https://fonts.googleapis.com/css2?family=Inconsolata:wght@200;300;400;500;600;700;800;900&display=swap" rel="stylesheet">
-		<link rel="stylesheet" href="styles.css">
+		<link rel="stylesheet" href="assets/css/register.css">
 	</head>
 	<body>
 		<div class="register">
@@ -28,86 +28,7 @@
         </div>
         
         <script src="https://code.jquery.com/jquery-3.4.1.min.js"></script>
-        <script src='assets/script.js'></script>        
+        <script src='assets/js/register.js'></script>
+        <script src='script.js'></script>      
 	</body>
 </html>
-<style>
-	body {
-    background-image: radial-gradient( circle at top left, darkgreen, black );
-    position: relative;
-    height: 100vh;
-    text-shadow: 0 0 2px limegreen;
-}
-.register {
-    margin: 0;
-      position: absolute;
-      top: 45%;
-      left: 50%;
-      -ms-transform: translate(-45%, -45%);
-      transform: translate(-50%, -50%);
-    width: 300px;
-    margin: 0 auto;
-    font-family: 'Inconsolata', monospace;
-    padding: 30px;
-    border: solid black 5px;
-}
-.register h1  {
-    text-align: center;
-    color: lime;
-    font-size: 24px;
-    padding: 0 0 20px 0;
-}
-.register input[type="password"],
-.register input[type="text"] {
-    width: 100%;
-    padding: 15px;
-    border: 1px solid lime;
-    margin-bottom: 15px;
-    box-sizing:border-box;
-    background-color: black;
-    color: lime;
-    caret-color: lime;
-}
-.register ::placeholder,
-.register ::placeholder {
-    color: lime;
-    opacity: 1;
-}
-.password p {
-    margin-top: 0;
-    display: hidden;
-    text-shadow: none;
-    color: darkgreen;
-}
-.password p:hover {
-    display: block;
-    margin-top: 0;
-    color: lime;
-}
-.show {
-    text-align: center;
-    color: lime;
-}
-#button {
-    width: 100%;
-    padding: 15px;
-    background-color: black;
-    border: 1px solid lime;
-    box-sizing: border-box;
-    cursor: pointer;
-    font-weight: bold;
-    color: lime;
-}
-.register #link {
-    text-align: center;
-    margin-top: 30px;
-}
-.register #link a {
-    color:lime;
-}
-::selection {
-    background-color: lime;
-    color: black;
-    text-shadow: none;
-  }
-</style>
